feat: add Chaotic Nixpkgs repository and switch to cachyos-lto kernel
This commit introduces the Chaotic Nixpkgs repository as a new input to our Nix flake, enabling access to a wider range of packages.  It also switches the kernel to `cachyos-lto` from this new repository.

Additionally, a `build.nix` file has been added to facilitate local builds.  This is currently disabled but can be enabled in the future.